fix: optimize sync procedure and fix some bug (#1145)
